requirements.txt
streamlit
transformers
diffusers
accelerate
safetensors
torch
torchvision
torchaudio
Pillow
gTTS
python-dotenv